@@ -156,9 +164,10 @@
     "del": "2.x",
     "devtron": "1.x",
     "dir-compare": "^1.4.0",
-    "electron": "1.7.11",
-    "electron-builder": "^13.11.1",
-    "electron-builder-squirrel-windows": "^20.4.0",
+    "electron": "2.0.0",
+    "electron-builder": "^20.13.2",
+    "electron-builder-lib": "^20.15.3",
+    "electron-builder-squirrel-windows": "^20.15.0",
     "electron-devtools-installer": "^2.2.3",
     "eslint": "3.x",
     "eslint-config-appium": "1.x",
@@ -197,7 +207,7 @@
     "npm": ">=3.x"
   },
   "engines": {
-    "node": ">=6.x <=7.x",
-    "npm": ">=3.x <=4.x"
+    "node": ">=9.x",
+    "npm": ">=5.x"
   }
 }
