@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
-import { Card, Icon, Button, Spin, Tooltip } from 'antd';
+import { Card, Icon, Button, Spin, Tooltip, Modal } from 'antd';
 import Screenshot from './Screenshot';
 import SelectedElement from './SelectedElement';
 import Source from './Source';
@@ -21,7 +21,7 @@ export default class Inspector extends Component {
     this.state = {};
   }
 
-  componentWillMount () {
+  componentDidMount () {
     const curHeight = window.innerHeight;
     const curWidth = window.innerWidth;
     const needsResize = (curHeight < MIN_HEIGHT) || (curWidth < MIN_WIDTH);
@@ -46,7 +46,8 @@ export default class Inspector extends Component {
   render () {
     const {screenshot, screenshotError, selectedElement = {},
       applyClientMethod, quitSession, isRecording, showRecord, startRecording,
-      pauseRecording, showLocatorTestModal, screenshotInteractionMode} = this.props;
+      pauseRecording, showLocatorTestModal, screenshotInteractionMode, 
+      showKeepAlivePrompt, keepSessionAlive} = this.props;
     const {path} = selectedElement;
 
     let main = <div className={InspectorStyles['inspector-main']}>
@@ -76,7 +77,7 @@ export default class Inspector extends Component {
          className={InspectorStyles['source-tree-card']}>
           <Source {...this.props} />
         </Card>
-        {this.container && <SourceScrollButtons container={this.container} />}
+        {this.container && <SourceScrollButtons {...this.props} container={this.container} />}
       </div>
       <div id='selectedElementContainer' className={`${InspectorStyles['source-tree-container']} ${InspectorStyles['element-detail-container']}`}>
         <Card
@@ -139,6 +140,16 @@ export default class Inspector extends Component {
     return <div className={InspectorStyles['inspector-container']}>
       {controls}
       {main}
+      <Modal 
+        title="Session Inactive"
+        visible={showKeepAlivePrompt}
+        onOk={() => keepSessionAlive()}
+        onCancel={() => quitSession()}
+        okText="Keep Session Running"
+        cancelText="Quit Session"
+      >
+        <p>Your session is about to expire</p>
+      </Modal>
     </div>;
   }
 }
