@@ -18,8 +18,6 @@ var server = null;
 var logWatcher = null;
 var batchedLogs = [];
 
-let sessionDrivers = {};
-
 let appiumHandlers = {};
 let logFile;
 
@@ -37,21 +35,13 @@ async function deleteLogfile () {
 /**
  * Kill session associated with session browser window
  */
-async function killSession (sessionWinID) {
-  let driver = sessionDrivers[sessionWinID];
-  if (driver) {
-    let sessionID;
-    try {
-      await driver.getSessionId();
-      if (!driver._isAttachedSession) {
-        await driver.quit();
-      }
-    } catch (e) {
-      console.log(`Couldn't close session: ${sessionID || 'unknown session ID'}`);
-    }
-    delete sessionDrivers[sessionWinID];
-    delete appiumHandlers[sessionWinID];
+async function killSession (sessionWinID, killedByUser=false) {
+  let handler = appiumHandlers[sessionWinID];
+  if (handler) {
+    await handler.close('', killedByUser);  
   }
+  
+  delete appiumHandlers[sessionWinID];
 }
 
 function connectStartServer (win) {
@@ -153,11 +143,17 @@ function connectCreateNewSessionWindow (win) {
   });
 }
 
+function connectClearLogFile () {
+  ipcMain.on('appium-clear-logfile', async (event, {logfilePath}) => {
+    await fs.writeFile(logfilePath, '');
+  });
+}
+
 export function createNewSessionWindow (win) {
   // Create and open the Browser Window
   let sessionWin = new BrowserWindow({
-    width: 920,
-    minWidth: 920,
+    width: 1080,
+    minWidth: 1080,
     height: 570,
     minHeight: 570,
     title: "Start Session",
@@ -181,19 +177,13 @@ export function createNewSessionWindow (win) {
   // When you close the session window, kill its associated Appium session (if there is one)
   let sessionID = sessionWin.webContents.id;
   sessionWin.on('closed', async () => {
-    const driver = sessionDrivers[sessionID];
-    if (driver) {
-      if (!driver._isAttachedSession) {
-        await driver.quit();
-      }
-      delete sessionDrivers[sessionID];
-    }
+    await killSession(sessionID);
     sessionWin = null;
   });
 
   // When the main window is closed, close the session window too
   win.once('closed', () => {
-    sessionWin = null;
+    sessionWin.close();
   });
 
   // If it's dev, go ahead and open up the dev tools automatically
@@ -219,12 +209,12 @@ function connectCreateNewSession () {
 
     try {
       // If there is an already active session, kill it. Limit one session per window.
-      if (sessionDrivers[event.sender.id]) {
+      if (appiumHandlers[event.sender.id]) {
         await killSession(event.sender);
       }
 
       // Create the driver and cache it by the sender ID
-      let driver = sessionDrivers[event.sender.id] = wd.promiseChainRemote({
+      let driver = wd.promiseChainRemote({
         hostname: host,
         port,
         path,
@@ -233,7 +223,7 @@ function connectCreateNewSession () {
         https,
       });
       driver.configureHttp({rejectUnauthorized, proxy});
-      appiumHandlers[event.sender.id] = new AppiumMethodHandler(driver);
+      const handler = appiumHandlers[event.sender.id] = new AppiumMethodHandler(driver, event.sender);
 
       // If we're just attaching to an existing session, do that and
       // short-circuit the rest of the logic
@@ -261,6 +251,12 @@ function connectCreateNewSession () {
       let p = driver.init(desiredCapabilities);
       event.sender.send('appium-new-session-successful');
       await p;
+
+      if (host !== '127.0.0.1' && host !== 'localhost') {
+        handler.runKeepAliveLoop();
+      }
+
+
       // we don't really support the web portion of apps for a number of
       // reasons, so pre-emptively ensure we're in native mode before doing the
       // rest of the inspector startup. Since some platforms might not implement
@@ -280,6 +276,12 @@ function connectCreateNewSession () {
 function connectRestartRecorder () {
   ipcMain.on('appium-restart-recorder', (evt) => {
     appiumHandlers[evt.sender.id].restart();
+  });
+}
+
+function connectKeepAlive () {
+  ipcMain.on('appium-keep-session-alive', (evt) => {
+    appiumHandlers[evt.sender.id].keepSessionAlive();
   });
 }
 
@@ -305,12 +307,13 @@ function connectClientMethodListener () {
 
     try {
       if (methodName === 'quit') {
-        await killSession(renderer.id);
+        await killSession(renderer.id, true);
         // when we've quit the session, there's no source/screenshot to send
         // back
         renderer.send('appium-client-command-response', {
           source: null,
           screenshot: null,
+          windowSize: null,
           uuid,
           result: null
         });
@@ -384,6 +387,8 @@ function initializeIpc (win) {
   connectGetSessionsListener();
   connectRestartRecorder();
   connectMoveToApplicationsFolder();
+  connectKeepAlive();
+  connectClearLogFile();
 
   setTimeout(checkNewUpdates, 10000);
 }
