@@ -25,9 +25,8 @@ export default class HighlighterRects extends Component {
     const screenshotEl = this.props.containerEl.querySelector('img');
 
     // now update scale ratio
-    const {x1, x2} = parseCoordinates(this.props.source.children[0].children[0]);
     this.setState({
-      scaleRatio: (x2 - x1) / screenshotEl.offsetWidth
+      scaleRatio: (this.props.windowSize.width / screenshotEl.offsetWidth)
     });
 
   }
@@ -82,7 +81,7 @@ export default class HighlighterRects extends Component {
     const {swipeStart, swipeEnd, clearSwipeAction, applyClientMethod} = this.props;
     await applyClientMethod({
       methodName: 'swipe',
-      args: [swipeStart.x, swipeStart.y, swipeEnd.x - swipeStart.x, swipeEnd.y - swipeStart.y],
+      args: [swipeStart.x, swipeStart.y, swipeEnd.x, swipeEnd.y],
     });
     clearSwipeAction();
   }
