@@ -11,7 +11,7 @@ describe('updateChecker', function () {
   let latestVersion;
 
   before(async function () {
-    const latestReleaseUrl = `https://api.github.com/repos/appium/appium-desktop/releases/latest`;
+    const latestReleaseUrl = `https://api.github.com/repos/appium/appium-desktop/releases/latest?access_token=${process.env.GITHUB_TOKEN}`;
     const res = JSON.parse(await request.get(latestReleaseUrl, {headers: {'user-agent': 'node.js'}}));
     latestVersion = res.name;
   });
